# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kayak Trails API",
        "description": "API documentation for kayak trails application. Complete API reference for managing kayak routes, weather data, and geographical information.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "api.wartkinurt.pl/api/v1",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/dashboard/users/{user}/roles": {
            "get": {
                "tags": [
                    "Dashboard - User Roles"
                ],
                "summary": "Pobierz role użytkownika",
                "operationId": "5cc30b00421c9904741aec82a384fdc5",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role użytkownika pobrane pomyślnie"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/users/{user}/assign-roles": {
            "post": {
                "tags": [
                    "Dashboard - User Roles"
                ],
                "summary": "Przypisz role użytkownikowi",
                "operationId": "913f13cb68fa9c0d1fc0be7f8c65b633",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role przypisane pomyślnie"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/users/{user}/revoke-roles": {
            "delete": {
                "tags": [
                    "Dashboard - User Roles"
                ],
                "summary": "Odbierz role użytkownikowi",
                "operationId": "2ded5247355c0b938a762d0b40808e60",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role odebrane pomyślnie"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/users/{user}/sync-roles": {
            "put": {
                "tags": [
                    "Dashboard - User Roles"
                ],
                "summary": "Zsynchronizuj role użytkownika (zastąp wszystkie)",
                "operationId": "f0c41f8143644e3345f91fb3969eef68",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role zsynchronizowane pomyślnie"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}